feat: add input composables
