+++ packages/vlossom/src/composables/index.ts
@@ -1,7 +1,7 @@
 export * from './color-scheme-composable';
-export * from './overlay-composable';
 export * from './input-composable';
 export * from './input-form-composable';
 export * from './input-messages-composable';
-export * from './input-rule-composable';
+export * from './input-rules-composable';
+export * from './overlay-composable';
 export * from './style-set-composable';
